requirements.txt
aiohttp==3.6.2
APScheduler==3.6.3
async-timeout==3.0.1
asyncpg==0.21.0
attrs==19.3.0
certifi==2020.6.20
chardet==3.0.4
discord.py==1.4.1
googletrans==3.0.0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2020.8.18
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
multidict==4.7.6
pip==20.2.2
pygame==1.9.6
python-dotenv==0.14.0
pytz==2020.1
requests==2.24.0
rfc3986==1.4.0
setuptools==49.6.0
six==1.15.0
sniffio==1.1.0
tzlocal==2.1
urllib3==1.25.10
yarl==1.5.1